refer to project via 'project' and not 'abi' in setup_program_test
