+++ packages/fuels-code-gen/src/program_bindings/abigen/abigen_target.rs
@@ -1,6 +1,9 @@
-use std::convert::TryFrom;
+use std::{convert::TryFrom, str::FromStr};
+
+use proc_macro2::Ident;
 
 use crate::{
+    error,
     error::{Error, Result},
     program_bindings::abi_types::FullProgramABI,
     utils::Source,
@@ -42,4 +45,30 @@ pub enum ProgramType {
     Script,
     Contract,
     Predicate,
+}
+impl FromStr for ProgramType {
+    type Err = Error;
+
+    fn from_str(string: &str) -> std::result::Result<Self, Self::Err> {
+        let program_type = match string {
+            "Script" => ProgramType::Script,
+            "Contract" => ProgramType::Contract,
+            "Predicate" => ProgramType::Predicate,
+            _ => return Err(error!("'{string}' cannot be converted into a `ParamType`.")),
+        };
+
+        Ok(program_type)
+    }
+}
+
+impl TryFrom<Ident> for ProgramType {
+    type Error = syn::Error;
+
+    fn try_from(ident: Ident) -> std::result::Result<Self, Self::Error> {
+        ident
+            .to_string()
+            .as_str()
+            .parse()
+            .map_err(|e| Self::Error::new(ident.span(), e))
+    }
 }
