+++ packages/fuels-macros/src/abigen/parsing.rs
@@ -1,8 +1,7 @@
 use fuels_code_gen::{AbigenTarget, ProgramType};
-use proc_macro2::Ident;
 use syn::{
     parse::{Parse, ParseStream},
-    Error, LitStr, Result as ParseResult,
+    LitStr, Result,
 };
 
 use crate::parse_utils::{Command, UniqueNameValues};
@@ -38,19 +37,19 @@ pub(crate) struct MacroAbigenTargets {
 }
 
 impl Parse for MacroAbigenTargets {
-    fn parse(input: ParseStream) -> ParseResult<Self> {
+    fn parse(input: ParseStream) -> Result<Self> {
         let targets = Command::parse_multiple(input)?
             .into_iter()
             .map(MacroAbigenTarget::new)
-            .collect::<Result<Vec<_>, _>>()?;
+            .collect::<Result<Vec<_>>>()?;
 
         Ok(Self { targets })
     }
 }
 
 impl MacroAbigenTarget {
-    pub fn new(command: Command) -> syn::Result<Self> {
-        let program_type = Self::parse_program_type(command.name)?;
+    pub fn new(command: Command) -> Result<Self> {
+        let program_type = command.name.try_into()?;
 
         let name_values = UniqueNameValues::new(command.contents)?;
         name_values.validate_has_no_other_names(&["name", "abi"])?;
@@ -63,17 +62,5 @@ impl MacroAbigenTarget {
             abi,
             program_type,
         })
-    }
-
-    fn parse_program_type(ident: Ident) -> ParseResult<ProgramType> {
-        match ident.to_string().as_ref() {
-            "Contract" => Ok(ProgramType::Contract),
-            "Script" => Ok(ProgramType::Script),
-            "Predicate" => Ok(ProgramType::Predicate),
-            _ => Err(Error::new_spanned(
-                ident,
-                "Unsupported program type. Expected: 'Contract', 'Script' or 'Predicate'",
-            )),
-        }
     }
 }
