+++ packages/fuels/tests/bindings.rs
@@ -21,7 +21,7 @@ async fn compile_bindings_from_contract_file() {
         Wallets("wallet"),
         Abigen(Contract(
             name = "SimpleContract",
-            abi = "packages/fuels/tests/bindings/simple_contract"
+            project = "packages/fuels/tests/bindings/simple_contract"
         )),
         Deploy(
             name = "simple_contract_instance",
@@ -759,11 +759,11 @@ async fn shared_types() -> Result<()> {
         Abigen(
             Contract(
                 name = "ContractA",
-                abi = "packages/fuels/tests/bindings/sharing_types/contract_a"
+                project = "packages/fuels/tests/bindings/sharing_types/contract_a"
             ),
             Contract(
                 name = "ContractB",
-                abi = "packages/fuels/tests/bindings/sharing_types/contract_b"
+                project = "packages/fuels/tests/bindings/sharing_types/contract_b"
             ),
         ),
         Deploy(
@@ -877,7 +877,7 @@ async fn type_paths_respected() -> Result<()> {
         Wallets("wallet"),
         Abigen(Contract(
             name = "ContractA",
-            abi = "packages/fuels/tests/bindings/type_paths"
+            project = "packages/fuels/tests/bindings/type_paths"
         )),
         Deploy(
             name = "contract_a_instance",
